// src/app/Models/restaurant.model.ts
export interface RestaurantInformation {
    restaurantId: string;
    name: string;
    address: string;
    cityid: string;
    cuisineid: string;
    rating: number;
    reviews: number;
    cuisine: string;
    city: string;
}

export interface Restaurant {
    restaurantId: string;
    name: string;
    address: string;
    cityid: string;
    cuisineid: string;
    rating: number;
    reviews: number;
}